requirements.txt
bcrypt==3.1.4
cffi==1.11.5
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.5
pycparser==2.18
PyJWT==1.6.4
six==1.11.0
Werkzeug==0.14.1
